preparing release 0.8.5

+++ vagrant_provisioning/install_kafka.sh
@@ -3,12 +3,15 @@
 KAFKA_VERSION=0.8.2.2
 SCALA_VERSION=2.10
 
-echo "Installing Apache Kafka ${KAFKA_VERSION}"
+if [ ! -d /opt/kafka ]; then
+    echo "Installing Apache Kafka ${KAFKA_VERSION}"
+    mirror=$(curl -sS https://www.apache.org/dyn/closer.cgi\?as_json\=1 | jq -r '.preferred')
+    url="${mirror}kafka/${KAFKA_VERSION}/kafka_${SCALA_VERSION}-${KAFKA_VERSION}.tgz"
+    curl -sS ${url} --output /tmp/kafka.tgz
 
-mirror=$(curl -sS https://www.apache.org/dyn/closer.cgi\?as_json\=1 | jq -r '.preferred')
-url="${mirror}kafka/${KAFKA_VERSION}/kafka_${SCALA_VERSION}-${KAFKA_VERSION}.tgz"
-curl -sS ${url} --output /tmp/kafka.tgz
-
-tar xzf /tmp/kafka.tgz -C /opt
-rm /tmp/kafka.tgz
-mv /opt/kafka_${SCALA_VERSION}-${KAFKA_VERSION} /opt/kafka
+    tar xzf /tmp/kafka.tgz -C /opt
+    rm /tmp/kafka.tgz
+    mv /opt/kafka_${SCALA_VERSION}-${KAFKA_VERSION} /opt/kafka
+else
+    echo "Apache Kafka ${KAFKA_VERSION} already installed"
+fi
